[safesnap] multisend transaction creation & code clean up
