+++ src/plugins/safeSnap/constants.ts
@@ -1,0 +1,80 @@
+export const ModuleAbi = [
+  // Events
+  'event ProposalQuestionCreated(bytes32 indexed questionId, string indexed proposalId)',
+
+  // Read functions
+  'function avatar() view returns (address)', // Reality Module
+  'function executor() view returns (address)', // Dao Module
+  'function oracle() view returns (address)',
+  'function questionCooldown() view returns (uint32)',
+  'function buildQuestion(string proposalId, bytes32[] txHashes) view returns (string)',
+  'function executedProposalTransactions(bytes32 questionHash, bytes32 txHash) view returns (bool)',
+  'function questionIds(bytes32 questionHash) view returns (bytes32)',
+  'function minimumBond() view returns (uint256)',
+
+  // Write functions
+  'function addProposal(string proposalId, bytes32[] txHashes)',
+  'function executeProposalWithIndex(string proposalId, bytes32[] txHashes, address to, uint256 value, bytes data, uint8 operation, uint256 txIndex)'
+];
+
+export const OracleAbi = [
+  // Events
+  `event LogNewAnswer(
+    bytes32 answer,
+    bytes32 indexed question_id,
+    bytes32 history_hash,
+    address indexed user,
+    uint256 bond,
+    uint256 ts,
+    bool is_commitment
+  )`,
+
+  // Read functions
+  'function resultFor(bytes32 question_id) view returns (bytes32)',
+  'function getFinalizeTS(bytes32 question_id) view returns (uint32)',
+  'function getBond(bytes32 question_id) view returns (uint256)',
+  'function getBestAnswer(bytes32 question_id) view returns (uint32)',
+  'function balanceOf(address) view returns (uint256)',
+  'function getHistoryHash(bytes32 question_id) view returns (bytes32)',
+  'function isFinalized(bytes32 question_id) view returns (bool)',
+  'function token() view returns (address)',
+
+  // Write functions
+  'function submitAnswer(bytes32 question_id, bytes32 answer, uint256 max_previous) external payable',
+  'function submitAnswerERC20(bytes32 question_id, bytes32 answer, uint256 max_previous, uint256 tokens) external',
+  `function claimMultipleAndWithdrawBalance(
+    bytes32[] question_ids,
+    uint256[] lengths,
+    bytes32[] hist_hashes,
+    address[] addrs,
+    uint256[] bonds,
+    bytes32[] answers
+  ) public`,
+  'function withdraw() public'
+];
+
+export const TokenAbi = [
+  //Read functions
+  'function balanceOf(address account) view returns (uint256)',
+  'function decimals() view returns (uint32)',
+  'function symbol() view returns (string)',
+  'function allowance(address owner, address spender) external view returns (uint256)',
+
+  // Write functions
+  'function approve(address spender, uint256 value) external returns (bool)'
+];
+
+export const START_BLOCKS = {
+  1: 6531147,
+  4: 3175028
+};
+
+export const EIP712_TYPES = {
+  Transaction: [
+    { name: 'to', type: 'address' },
+    { name: 'value', type: 'uint256' },
+    { name: 'data', type: 'bytes' },
+    { name: 'operation', type: 'uint8' },
+    { name: 'nonce', type: 'uint256' }
+  ]
+};
