+++ src/plugins/safeSnap/index.ts
@@ -1,275 +1,33 @@
 import { isAddress } from '@ethersproject/address';
-import {
-  BigNumber,
-  isBigNumberish
-} from '@ethersproject/bignumber/lib/bignumber';
+import { BigNumber } from '@ethersproject/bignumber';
+import { isBigNumberish } from '@ethersproject/bignumber/lib/bignumber';
 import { isHexString } from '@ethersproject/bytes';
-import { HashZero } from '@ethersproject/constants';
 import { StaticJsonRpcProvider } from '@ethersproject/providers';
 import { keccak256 as solidityKeccak256 } from '@ethersproject/solidity';
-import { call, multicall, sendTransaction } from '../../utils';
-import { getProvider } from '../../utils';
+import { call, getProvider, multicall, sendTransaction } from '../../utils';
 import { _TypedDataEncoder } from '@ethersproject/hash';
 import { Contract } from '@ethersproject/contracts';
 import { Result } from '@ethersproject/abi';
-
-const EIP712_TYPES = {
-  Transaction: [
-    {
-      name: 'to',
-      type: 'address'
-    },
-    {
-      name: 'value',
-      type: 'uint256'
-    },
-    {
-      name: 'data',
-      type: 'bytes'
-    },
-    {
-      name: 'operation',
-      type: 'uint8'
-    },
-    {
-      name: 'nonce',
-      type: 'uint256'
-    }
-  ]
-};
-
-const ModuleAbi = [
-  // Events
-  'event ProposalQuestionCreated(bytes32 indexed questionId, string indexed proposalId)',
-
-  // Read functions
-  'function avatar() view returns (address)', // Reality Module
-  'function executor() view returns (address)', // Dao Module
-  'function oracle() view returns (address)',
-  'function questionCooldown() view returns (uint32)',
-  'function buildQuestion(string proposalId, bytes32[] txHashes) view returns (string)',
-  'function executedProposalTransactions(bytes32 questionHash, bytes32 txHash) view returns (bool)',
-  'function questionIds(bytes32 questionHash) view returns (bytes32)',
-  'function minimumBond() view returns (uint256)',
-
-  // Write functions
-  'function addProposal(string proposalId, bytes32[] txHashes)',
-  'function executeProposalWithIndex(string proposalId, bytes32[] txHashes, address to, uint256 value, bytes data, uint8 operation, uint256 txIndex)'
-];
-
-const OracleAbi = [
-  // Events
-  `event LogNewAnswer(
-    bytes32 answer,
-    bytes32 indexed question_id,
-    bytes32 history_hash,
-    address indexed user,
-    uint256 bond,
-    uint256 ts,
-    bool is_commitment
-  )`,
-
-  // Read functions
-  'function resultFor(bytes32 question_id) view returns (bytes32)',
-  'function getFinalizeTS(bytes32 question_id) view returns (uint32)',
-  'function getBond(bytes32 question_id) view returns (uint256)',
-  'function getBestAnswer(bytes32 question_id) view returns (uint32)',
-  'function balanceOf(address) view returns (uint256)',
-  'function getHistoryHash(bytes32 question_id) view returns (bytes32)',
-  'function isFinalized(bytes32 question_id) view returns (bool)',
-  'function token() view returns (address)',
-
-  // Write functions
-  'function submitAnswer(bytes32 question_id, bytes32 answer, uint256 max_previous) external payable',
-  'function submitAnswerERC20(bytes32 question_id, bytes32 answer, uint256 max_previous, uint256 tokens) external',
-  `function claimMultipleAndWithdrawBalance(
-    bytes32[] question_ids,
-    uint256[] lengths,
-    bytes32[] hist_hashes,
-    address[] addrs,
-    uint256[] bonds,
-    bytes32[] answers
-  ) public`,
-  'function withdraw() public'
-];
-
-const TokenAbi = [
-  //Read functions
-  'function balanceOf(address account) view returns (uint256)',
-  'function decimals() view returns (uint32)',
-  'function symbol() view returns (string)',
-  'function allowance(address owner, address spender) external view returns (uint256)',
-
-  // Write functions
-  'function approve(address spender, uint256 value) external returns (bool)'
-];
-
-const START_BLOCKS = {
-  1: 6531147,
-  4: 3175028
-};
-export interface ModuleTransaction {
-  to: string;
-  value: string;
-  data: string;
-  operation: string;
-  nonce: string;
-}
-
-export interface ProposalDetails {
-  dao: string;
-  oracle: string;
-  cooldown: number;
-  proposalId: string;
-  questionId: string | undefined;
-  executionApproved: boolean;
-  finalizedAt: number | undefined;
-  nextTxIndex: number | undefined;
-  transactions: ModuleTransaction[];
-  txHashes: string[];
-  currentBond: BigNumber | undefined;
-  isApproved: boolean;
-  endTime: number | undefined;
-}
-
-const buildQuestion = async (proposalId: string, txHashes: string[]) => {
-  const hashesHash = solidityKeccak256(['bytes32[]'], [txHashes]).slice(2);
-  return `${proposalId}␟${hashesHash}`;
-};
-
-const getProposalDetails = async (
-  provider: StaticJsonRpcProvider,
-  network: string,
-  moduleAddress: string,
-  questionHash: string,
-  txHashes: string[]
-): Promise<{ questionId: string; nextTxIndex: number | undefined }> => {
-  const proposalInfo = (
-    await multicall(
-      network,
-      provider,
-      ModuleAbi,
-      [[moduleAddress, 'questionIds', [questionHash]]].concat(
-        txHashes.map((txHash) => [
-          moduleAddress,
-          'executedProposalTransactions',
-          [questionHash, txHash]
-        ])
-      )
-    )
-  ).map((res) => res[0]);
-  const questionId = proposalInfo[0];
-  // We need to offset the index by -1 the first element is the questionId
-  const nextIndexToExecute = proposalInfo.indexOf(false, 1) - 1;
-  return {
-    questionId: questionId !== HashZero ? questionId : undefined,
-    nextTxIndex:
-      nextIndexToExecute < 0 || nextIndexToExecute >= txHashes.length
-        ? undefined
-        : nextIndexToExecute
-  };
-};
-
-const getModuleDetails = async (
-  provider: StaticJsonRpcProvider,
-  network: string,
-  moduleAddress: string
-): Promise<{
-  dao: string;
-  oracle: string;
-  cooldown: number;
-  minimumBond: number;
-}> => {
-  let moduleDetails;
-  try {
-    // Assume module is Reality Module
-    moduleDetails = await multicall(network, provider, ModuleAbi, [
-      [moduleAddress, 'avatar'],
-      [moduleAddress, 'oracle'],
-      [moduleAddress, 'questionCooldown'],
-      [moduleAddress, 'minimumBond']
-    ]);
-  } catch (err) {
-    // The Reality Module doesn't have an avatar field, causing tx to fails.
-    // Assume module is Dao Module (old version)
-    moduleDetails = await multicall(network, provider, ModuleAbi, [
-      [moduleAddress, 'executor'],
-      [moduleAddress, 'oracle'],
-      [moduleAddress, 'questionCooldown'],
-      [moduleAddress, 'minimumBond']
-    ]);
-  }
-
-  return {
-    dao: moduleDetails[0][0],
-    oracle: moduleDetails[1][0],
-    cooldown: moduleDetails[2][0],
-    minimumBond: moduleDetails[3][0]
-  };
-};
-
-const checkPossibleExecution = async (
-  provider: StaticJsonRpcProvider,
-  network: string,
-  oracleAddress: string,
-  questionId: string | undefined
-): Promise<{
-  executionApproved: boolean;
-  finalizedAt: number | undefined;
-}> => {
-  if (questionId) {
-    try {
-      const result = await multicall(network, provider, OracleAbi, [
-        [oracleAddress, 'resultFor', [questionId]],
-        [oracleAddress, 'getFinalizeTS', [questionId]]
-      ]);
-
-      return {
-        executionApproved: BigNumber.from(result[0][0]).eq(BigNumber.from(1)),
-        finalizedAt: BigNumber.from(result[1][0]).toNumber()
-      };
-    } catch (e) {
-      // We expect an error while the question is not answered yet
-    }
-  }
-  return {
-    executionApproved: false,
-    finalizedAt: undefined
-  };
-};
-
-const retrieveInfoFromOracle = async (
-  provider: StaticJsonRpcProvider,
-  network: string,
-  oracleAddress: string,
-  questionId: string | undefined
-): Promise<{
-  currentBond: BigNumber | undefined;
-  isApproved: boolean;
-  endTime: number | undefined;
-}> => {
-  if (questionId) {
-    const result = await multicall(network, provider, OracleAbi, [
-      [oracleAddress, 'getFinalizeTS', [questionId]],
-      [oracleAddress, 'getBond', [questionId]],
-      [oracleAddress, 'getBestAnswer', [questionId]]
-    ]);
-
-    const currentBond = BigNumber.from(result[1][0]);
-    const answer = BigNumber.from(result[2][0]);
-
-    return {
-      currentBond,
-      isApproved: answer.eq(BigNumber.from(1)),
-      endTime: BigNumber.from(result[0][0]).toNumber()
-    };
-  }
-  return {
-    currentBond: undefined,
-    isApproved: false,
-    endTime: undefined
-  };
-};
+import { ModuleTransaction, ProposalDetails } from './models';
+import {
+  EIP712_TYPES,
+  ModuleAbi,
+  OracleAbi,
+  START_BLOCKS,
+  TokenAbi
+} from './constants';
+import {
+  buildQuestion,
+  checkPossibleExecution,
+  getModuleDetails,
+  getProposalDetails
+} from './utils/realityModule';
+import { retrieveInfoFromOracle } from './utils/realityETH';
+import {
+  createMultiSendTx,
+  getMultiSend,
+  MULTI_SEND_VERSION
+} from './utils/multiSend';
 
 export default class Plugin {
   public author = 'Gnosis';
@@ -277,6 +35,21 @@ export default class Plugin {
   public name = 'SafeSnap';
   public website = 'https://safe.gnosis.io';
   public options: any;
+
+  static createMultiSendTx(
+    txs: ModuleTransaction[],
+    nonce: number,
+    multiSendAddress: string
+  ) {
+    return createMultiSendTx(txs, nonce, multiSendAddress);
+  }
+
+  static getMultiSend(
+    network: number | string,
+    version: MULTI_SEND_VERSION = MULTI_SEND_VERSION.V1_3_0
+  ) {
+    return getMultiSend(network, version);
+  }
 
   validateTransaction(transaction: ModuleTransaction) {
     const addressEmptyOrValidate =
@@ -313,12 +86,11 @@ export default class Plugin {
       verifyingContract: moduleAddress
     };
     return transactions.map((tx) => {
-      const txHash = _TypedDataEncoder.hash(domain, EIP712_TYPES, {
+      return _TypedDataEncoder.hash(domain, EIP712_TYPES, {
         ...tx,
         nonce: tx.nonce || '0',
         data: tx.data || '0x'
       });
-      return txHash;
     });
   }
