+++ src/plugins/safeSnap/utils/realityETH.ts
@@ -1,0 +1,37 @@
+import { StaticJsonRpcProvider } from '@ethersproject/providers';
+import { BigNumber } from '@ethersproject/bignumber';
+import { multicall } from '../../../utils';
+import { OracleAbi } from '../constants';
+
+export const retrieveInfoFromOracle = async (
+  provider: StaticJsonRpcProvider,
+  network: string,
+  oracleAddress: string,
+  questionId: string | undefined
+): Promise<{
+  currentBond: BigNumber | undefined;
+  isApproved: boolean;
+  endTime: number | undefined;
+}> => {
+  if (questionId) {
+    const result = await multicall(network, provider, OracleAbi, [
+      [oracleAddress, 'getFinalizeTS', [questionId]],
+      [oracleAddress, 'getBond', [questionId]],
+      [oracleAddress, 'getBestAnswer', [questionId]]
+    ]);
+
+    const currentBond = BigNumber.from(result[1][0]);
+    const answer = BigNumber.from(result[2][0]);
+
+    return {
+      currentBond,
+      isApproved: answer.eq(BigNumber.from(1)),
+      endTime: BigNumber.from(result[0][0]).toNumber()
+    };
+  }
+  return {
+    currentBond: undefined,
+    isApproved: false,
+    endTime: undefined
+  };
+};
